Add files via upload
When I first booted my raspberry pi for the first time i couldn't get the display to stop spilling way off of the page, I messed with the "config.txt" for about a week before i finally got it right. These files will be a collaberation of actual configuration set-ups from real people. The aim here is to make the whole display issue a bit easier to diagnose, and fix. 

-Please commit your config.txt file for research purposes. Include your:  
          Screen size
          Screen type
          Full copy of config.txt file
   If there is anything else that you think will be useful/helpful to new raspi users then feel free to branch or add to this repo.